🚚 Moving files to server dir
